[ticket/15977][ci skip] Moving to controllers: MCP - controllers directory
PHPBB3-15977
