+++ phpBB/phpbb/mcp/controller/logs.php
@@ -11,7 +11,7 @@
 *
 */
 
-namespace phpbb\mcp;
+namespace phpbb\mcp\controller;
 
 class logs
 {
