+++ phpBB/phpbb/mcp/controller/reports.php
@@ -11,7 +11,7 @@
 *
 */
 
-namespace phpbb\mcp;
+namespace phpbb\mcp\controller;
 
 class reports
 {
